test for elapsed time

+++ src/timer/Clock.java
@@ -2,5 +2,5 @@ package timer;
 
 public interface Clock {
 
-	long currentTimeSeconds();
+	int currentTimeSeconds();
 }
