+++ src/timer/SystemClock.java
@@ -3,8 +3,8 @@ package timer;
 public class SystemClock implements Clock {
 
 	@Override
-	public long currentTimeSeconds() {
-		return System.currentTimeMillis() / 1000;
+	public int currentTimeSeconds() {
+		return (int) (System.currentTimeMillis() / 1000);
 	}
 	
 }
